refactor: home component separated
