+++ src/TechStack/index.js
@@ -1,0 +1,2 @@
+import TechStack from './TechStack'
+export default TechStack
