+++ src/components/AuthorInfo/AuthorInfo.js
