+++ src/components/AuthorInfo/AuthorInfo.style.js
