+++ src/pages/index.js
@@ -1,14 +1,16 @@
-import React from 'react';
-
 import Layout from '@components/Layout';
 import SEO from '@components/Seo';
-import AuthorInfo from '@src/components/Home';
+import AuthorInfo from '@src/components/AuthorInfo';
+import TechStack from '@src/TechStack';
+import React from 'react';
+
 
 export default function IndexPage() {
   return (
     <Layout>
       <SEO title={`Home`} />
       <AuthorInfo />
+      <TechStack />
     </Layout>
   );
 }
